replaced input/output with read/write naming and started making reader interface capable of having a reader per topic(s) which leads to topic(partition) routing

+++ src/Kafka.Common/Model/CompressionType.cs
@@ -1,6 +1,6 @@
 ﻿namespace Kafka.Common.Model
 {
-    public enum CompressionType : short
+    public enum CompressionType : int
     {
         None = 0,
         Gzip = 1,
