+++ src/Kafka.Common/Model/CoordinatorType.cs
@@ -1,6 +1,6 @@
 ﻿namespace Kafka.Common.Model
 {
-    public enum CoordinatorType : sbyte
+    public enum CoordinatorType : int
     {
         GROUP = 0,
         TRANSACTION =1
